Loja aleatoria

+++ Assets/_Scripts/U_Variables/UniversalVariables.cs
@@ -6,7 +6,7 @@ namespace _Scripts.U_Variables
     public class UniversalVariables : MonoSingleton<UniversalVariables>
     {
         //Variavél de dinheiro para a loja e venda das curas para os pacientes
-        public static int Money { get; set; } = 100;
+        public static float Money { get; set; } = 100;
         public static int Reputation { get; set; } = 30;
         public static float SocialAlignment { get; set; } = 50; // pra perto de 0 esta alinhado aos humildes mais perto de 100 aos ricos
         
@@ -15,7 +15,7 @@ namespace _Scripts.U_Variables
             DontDestroyOnLoad(this.gameObject);
         }
 
-        public void ModifyMoney(int amount, bool add)
+        public void ModifyMoney(float amount, bool add)
         {
             if (add)
                 Money += amount;
